Get rid of accumulator in `spineTo`
The algorithm implemented in `spineTo` is trivially recursive and the
value it returns is inherently a lazy one (a list), so it makes no sense
to implement it as a left fold, since it could be only useful if we
could avoid the buildup of the result value.

Remove all the complexity by rewriting it in a equivalent and very
direct form.

Also, get rid of the `toList` call, since `concatMap` works on any
Foldable.