[LIGO-382] Refactor make a bit
Problem: `make` sometimes throws a very annoying error.

Also, it always annoyed me that it expects a tuple and not a function...

Solution: Prefix `make` with `unsafe` and make a safe version and a not
so safe version of it. Make everything a function with 2 arguments
instead of expecting a tuple.