[LIGO-405] Reuse exceptions instead of transformers
Problem: There is a regression caused with LIGO-382: state is not
restored properly. This is because in the past we just threw an
exception to discard state from State, but now we use the monadic bind
with a ExceptT stack. Furthermore, we use this from `IO` anyway, so
having to handle both `ExceptT` and `IO` is a smell, so revert most of
what was done in commit 1cf4838.

Solution: Use old definitions of functions, throwing `HandlerFailed`.